requirements.txt
streamlit
spacy
openai-whisper
torch
ffmpeg-python

# streamlit run app.py